Support CREA metadata resolving
It doesn't appear that CREA's representation of RETS follows the
convention of having a "Lookup" string fragment in fields that can be
resolved using the lookup table.

This modification supports the original lookup method, but falls back to
a plain check to determine if the lookup types has a key that matches
the name of the field.

If it does, then we'll use that field to do the lookup and resolve.